Restructure the UI. Splitting up the GameBoard.
The GameBoard got split up:
    PlayerStage is now the main stage.
    GameBoard is a Pane and is responsable for player-specific
        interactions. it is composed of a DicePane and a
        ChooseCategoryPane.
    DicePane uses DieButton's.
    FooterPane displays game status information and is shared by all
        PlayerStage's.

The event classes now implement the Event(eventType) constructor,
instead of the Event(source, target, eventType) contructor.
The previous implementation was incorrectly using the Events API.

SetAsideDieEvent's getter got renamed to the more sensible getDie().

The Controller has suffered some spaghettification trying to cope
with the changes, but is hanging in there for the time beeing.